+++ plugins/wme-sitebuilder/CHANGELOG.md
@@ -5,3 +5,8 @@
 ### Minor Changes
 
 - fcadf1d: Consolidate the WME Sitebuilder WordPress plugin and library into the same repository with other Sitebuilder packages.
+
+### Patch Changes
+
+- Updated dependencies [0413284]
+  - @moderntribe/sitebuilder@1.7.0